Button updates, naming updates, cleared old unused button images
