+++ app/src/main/java/com/example/wrinklethinkle/view/InsideHouseFragment.kt
@@ -6,21 +6,27 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.wrinklethinkle.R
+import com.example.wrinklethinkle.databinding.FragmentInsideHouseBinding
+import androidx.navigation.fragment.findNavController
 
 class InsideHouseFragment : Fragment() {
+    private var insideHouseFragmentBinding: FragmentInsideHouseBinding? = null
+    private val binding get() = insideHouseFragmentBinding!!
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        insideHouseFragmentBinding = FragmentInsideHouseBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_inside_house, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.GrowButton.setOnClickListener {
+            findNavController().navigate(R.id.action_InsideHouseFragment_to_GrowFragment)
+        }
     }
 
-
+    override fun onDestroyView() {
+        super.onDestroyView()
+        insideHouseFragmentBinding = null
+    }
 }
