+++ app/src/main/res/layout/grow_fragment.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.FlowerClickFragment"
+        tools:context=".view.GrowFragment"
         android:id="@+id/constraintLayout">
 
     <ImageView
