+++ app/src/main/res/values/strings.xml
@@ -27,7 +27,7 @@
     <!-- Image content descriptions -->
     <string name="app_start_background_cd">app_start_bg_cd</string>
     <string name="go_home_button_cd">go_home_btn_cd</string>
-    <string name="flower_click_fragment_label">flower click frag</string>
+    <string name="grow_fragment_label">Grow Fragment</string>
     <string name="app_start_fragment_label">app start frag</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
